Refactors app organization to use 'page' concept
The application is now organized in pages: components will no longer
render independently to form page content based on the current URL.
Instead, components are grouped and displayed through wrapping page
components.

This strategy allows to better control which components are rendered per
URL and allows the use of the `switch` routing component to have a
redirect to the main page when the user navigates to an uncontrolled
route.