Extracts geolocation from CurrentWeather
The geolocation logic from the `CurrentWeather` component is extracted
into it's own component `GeoHandler`.

This was necessary as `CurrentWeather` was the only component to trigger
the necessary steps to fetch data, making other components like
`Forecast` somewhat dependent on it.

`GeoHandler` is now rendered by the main `App` component outside of the
router configuration, thus ensuring that the geolocation step and data
fetching is always triggered.