Adds settings navigation functionality to Toolbar
`Toolbar` can now navigate to the `/settings` route. This is not
achieved with a `Link` component because `Toolbar` is declared outside
the router in the `App` component. The latter was chosen because Toolbar
should always be displayed regardless of the selected route. Instead, it
dispatches a push action to the router since it is now connected to
Redux.