Removes branch restriction in CircleCI config
CircleCI was configured to run only for the master branch in order to
avoid deploying to Netlify from other branches. However, this solution
also blocked CircleCI from checking pull requests originating from other
branches.

This commit removes the attribute that limited CircleCI to run only on
the master branch and instead added a check in the deploy step to check
if the branch is master (and deploy) or not.