Fixes Toolbar not displaying without API data
`Toolbar` was dependent on weather data to be available only because on
the main page it would show the location name.

However, this meant that it would not render anything until data is
fetched. This is not desirable because settings wouldn't be reachable
and, if opening the app from the settings or license routes, the
application would display as expected except for the toolbar until data
is pulled.

This commit fixes this issue by not returning `null` if there is no
data.