Implements CircleCI Workflows
Divides the unit and e2e tests into separate jobs so that they can run
in parallel. Also, separate the deploy section into a standalone job
that will run only if test jobs are successful and if the branch is
master.

In order to achieve the separation of the two test jobs, the Jest
configuration was split into two configuration files.