Disables zoom
By adding the property `user-scalable=no` in the `viewport` meta tag,
user zoom is disabled. This behaviour is desired so that the application
resembles more a native mobile application.

This does not affect much accessability as the application is designed
to be responsive and does react well to font and UI OS scaling settings.